SensorOverlay documentation
