+++ src/components/SensorOverlay/components/DraggablePoint.test.tsx
