+++ src/components/SensorOverlay/components/DraggablePoint.tsx
@@ -6,7 +6,7 @@ import {
   ConnectDragSource,
   ConnectDragPreview,
 } from 'react-dnd';
-import { DragTargets } from './constants';
+import { DragTargets } from '../constants';
 
 const boxSource = {
   beginDrag(props: DraggablePointProps) {
