+++ src/components/SensorOverlay/components/SvgLine.tsx
